requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.14
googleapis-common-protos==1.65.0
idna==3.10
importlib_metadata==8.4.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
protobuf==4.25.5
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
